# packages/shared/openapi/healthpick.openapi.json
{
  "components": {
    "schemas": {
      "ChatFinalEvent": {
        "additionalProperties": false,
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "title": "Citations",
            "type": "array"
          },
          "model": {
            "$ref": "#/components/schemas/ModelInfo"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "route": {
            "enum": [
              "nutrition",
              "platform",
              "mixed",
              "out_of_scope"
            ],
            "title": "Route",
            "type": "string"
          },
          "safety": {
            "$ref": "#/components/schemas/SafetyResult"
          }
        },
        "required": [
          "request_id",
          "answer",
          "route",
          "citations",
          "safety",
          "model"
        ],
        "title": "ChatFinalEvent",
        "type": "object"
      },
      "ChatRequest": {
        "additionalProperties": false,
        "properties": {
          "conversation_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Conversation Id",
            "type": "string"
          },
          "message": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "profile_patch": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProfilePatch"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "conversation_id",
          "message"
        ],
        "title": "ChatRequest",
        "type": "object"
      },
      "Citation": {
        "additionalProperties": false,
        "properties": {
          "chunk_id": {
            "title": "Chunk Id",
            "type": "string"
          },
          "excerpt": {
            "title": "Excerpt",
            "type": "string"
          },
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "section": {
            "title": "Section",
            "type": "string"
          },
          "source": {
            "enum": [
              "A",
              "B",
              "C"
            ],
            "title": "Source",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "source",
          "title",
          "section",
          "page",
          "chunk_id",
          "excerpt"
        ],
        "title": "Citation",
        "type": "object"
      },
      "ComponentStatus": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "status": {
            "enum": [
              "configured",
              "mock",
              "not_configured",
              "not_checked",
              "disabled"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "ComponentStatus",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "request_id",
          "code",
          "message",
          "retryable"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "HealthResponse": {
        "additionalProperties": false,
        "properties": {
          "dependencies": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ComponentStatus"
            },
            "title": "Dependencies",
            "type": "object"
          },
          "environment": {
            "enum": [
              "development",
              "test",
              "production"
            ],
            "title": "Environment",
            "type": "string"
          },
          "modes": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Modes",
            "type": "object"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "service": {
            "const": "healthpick-api",
            "default": "healthpick-api",
            "title": "Service",
            "type": "string"
          },
          "status": {
            "const": "ok",
            "default": "ok",
            "title": "Status",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "version",
          "environment",
          "request_id",
          "modes",
          "dependencies"
        ],
        "title": "HealthResponse",
        "type": "object"
      },
      "ModelInfo": {
        "additionalProperties": false,
        "properties": {
          "mode": {
            "enum": [
              "real",
              "mock"
            ],
            "title": "Mode",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "name",
          "mode"
        ],
        "title": "ModelInfo",
        "type": "object"
      },
      "ProfilePatch": {
        "additionalProperties": false,
        "properties": {
          "allergies": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Allergies",
            "type": "array"
          },
          "conditions": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Conditions",
            "type": "array"
          },
          "disliked_foods": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Disliked Foods",
            "type": "array"
          },
          "goal": {
            "anyOf": [
              {
                "enum": [
                  "fat_loss",
                  "muscle_gain",
                  "stable_glucose"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Goal"
          }
        },
        "title": "ProfilePatch",
        "type": "object"
      },
      "SafetyResult": {
        "additionalProperties": false,
        "properties": {
          "matched_rules": {
            "items": {
              "type": "string"
            },
            "title": "Matched Rules",
            "type": "array"
          },
          "required_notice": {
            "enum": [
              "general",
              "professional",
              "emergency"
            ],
            "title": "Required Notice",
            "type": "string"
          },
          "risk_level": {
            "enum": [
              "S0",
              "S1",
              "S2",
              "S3"
            ],
            "title": "Risk Level",
            "type": "string"
          }
        },
        "required": [
          "risk_level",
          "required_notice"
        ],
        "title": "SafetyResult",
        "type": "object"
      }
    }
  },
  "info": {
    "summary": "Traceable nutrition Q&A with hard source isolation.",
    "title": "HealthPick AI Nutrition Advisor API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/healthz": {
      "get": {
        "operationId": "healthz_healthz_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server error"
          }
        },
        "summary": "Healthz",
        "tags": [
          "system"
        ]
      }
    },
    "/v1/chat/stream": {
      "post": {
        "description": "SSE event order: meta, zero or more citation events, token events, then final. Failures detected before streaming use the stable JSON ErrorResponse contract.",
        "operationId": "chat_stream_v1_chat_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "description": "meta, citation, token and final SSE events",
                  "type": "string"
                }
              }
            },
            "description": "Ordered SSE stream ending with a ChatFinalEvent payload"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Provider or evidence failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Provider not configured or unavailable"
          }
        },
        "summary": "Stream one evidence-grounded chat response",
        "tags": [
          "chat"
        ]
      }
    }
  }
}